+++ CHANGELOG.md
@@ -8,6 +8,8 @@ This change log follows the conventions of
 
 ### Added
 
+- We can now obtain track metadata including the waveform and beat grid even for tracks that have not been analyzed by rekordbox when talking to a CDJ-3000 and using a real player number.
+  This means that such tracks can interact usefully with almost all Beat Link features.
 - The ability to provide limited features with the Opus Quad, even though that hardware does not really support the Pro DJ Link protocol, by leveraging its ability to work with rekordbox lighting. Thanks to [@cprepos](https://github.com/cprepos) for doing most of this work! Note that because the Opus Quad does not send beat packets, we might be up to 200ms out of sync with beats, since that is how often status packets arrive.
 - Support for the three-band waveform style introduced with the CDJ-3000.
 - The ability to proxy metadata from mounted archive files corresponding to USB media mounted in the Opus Quad, which cannot itself provide that information.